requirements.txt
python-dotenv
streamlit
openai
google-generativeai
loguru
PyPDF2